refactor(slack): extract the user/channel directory into user-directory.ts
normalize.ts had grown to ~1,400 lines mixing three concerns. Split the
first one out: the Slack user/channel resolution + caching layer
(resolveSlackUser / resolveSlackChannel / resolveSlackUserSync, the LRU
caches, in-flight de-dup, and the test-only cache helpers) moves to a
dedicated user-directory.ts. This is the directory that reads Slack's
authenticated Web API (users.info / conversations.info), distinct from the
untrusted-ingress normalization the rest of the file does.

Pure move: normalize.ts imports SlackUserInfo + resolveSlackUserSync back
from the new module (SlackUserActorFields stays with the actor helpers);
socket-mode and the tests import the resolvers/cache helpers from
user-directory.js. No behavior change.
